satopon_1/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
python-dotenv
python-jose
email-validator
redis>=4.4.0
google-auth>=2.40.3
requests


